feat(CMP-5384): Moved content reference from media to content/models
Wanting review on any helper functions a content reference would benefit from, maybe being able to request the reference?
