+++ src/index.ts
@@ -4,6 +4,7 @@ export * from './lib/ContentClientConfig';
 export * from './lib/content/model/ContentBody';
 export * from './lib/content/model/ContentItem';
 export * from './lib/content/model/ContentLifecycle';
+export * from './lib/content/model/ContentReference';
 export * from './lib/content/model/ContentMeta';
 export * from './lib/content/model/Edition';
 export * from './lib/content/model/FragmentMeta';
@@ -14,6 +15,5 @@ export * from './lib/media/Image';
 export * from './lib/media/ImageUrlBuilder';
 export * from './lib/media/model/ImageFormat';
 export * from './lib/media/Video';
-export * from './lib/media/ContentReference';
 
 export * from './lib/rendering/model/RenderedContentItem';
